deal with unknown content length
git-svn-id: https://svn.apache.org/repos/asf/incubator/ivy/trunk@484395 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/fr/jayasoft/ivy/url/BasicURLHandler.java
@@ -118,7 +118,7 @@ public class BasicURLHandler extends AbstractURLHandler {
             srcConn = src.openConnection();
             int contentLength = srcConn.getContentLength();
             FileUtil.copy(srcConn.getInputStream(), dest, l);
-            if (dest.length() != contentLength) {
+            if (dest.length() != contentLength && contentLength != -1) {
             	dest.delete();
             	throw new IOException("Downloaded file size doesn't match expected Content Length for "+src+". Please retry.");
             }
